+++ cmd/root/cli/cli.go
@@ -18,6 +18,8 @@ package cli
 
 import (
 	"bufio"
+	"bytes"
+	"errors"
 	"fmt"
 	"os/exec"
 	"strings"
@@ -29,7 +31,7 @@ import (
 
 type CmdExecutor interface {
 	Execute(mask bool, args ...string) (out string, err error)
-	ExecuteAndStreamOutput(mask bool, outputHandler OutputHandler, args ...string) error
+	ExecuteAndStreamOutput(mask bool, outputHandler OutputHandler, pat string, args ...string) error
 }
 
 type UnableToStartErr struct {
@@ -62,7 +64,10 @@ type Cli struct {
 // If the command runs successfully, the function will return nil. If there's an error executing the command,
 // it will return an error. Note that an error from the command itself (e.g., a non-zero exit status) will also
 // be returned as an error from this function.
-func (c *Cli) ExecuteAndStreamOutput(mask bool, outputHandler OutputHandler, args ...string) error {
+func (c *Cli) ExecuteAndStreamOutput(mask bool, outputHandler OutputHandler, pat string, args ...string) error {
+	if len(args) == 0 {
+		return errors.New("must have an argument but none was present")
+	}
 	// Log the command that's about to be run
 	logArgs(mask, args)
 
@@ -83,13 +88,20 @@ func (c *Cli) ExecuteAndStreamOutput(mask bool, outputHandler OutputHandler, arg
 	}
 	stdErrScanner := bufio.NewScanner(stderr)
 
+	if pat != "" {
+		buff := bytes.Buffer{}
+		_, err := buff.WriteString(pat)
+		if err != nil {
+			return err
+		}
+		cmd.Stdin = &buff
+	}
 	// Start the command
 	if err := cmd.Start(); err != nil {
 		return UnableToStartErr{Err: err, Cmd: strings.Join(args, " ")}
 	}
 	var wg sync.WaitGroup
 	wg.Add(2)
-
 	// Asynchronously read the output from the command line by line
 	// and pass it to the outputHandler. This runs in a goroutine
 	// so that we can also read the error output at the same time.